requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
python-jose[cryptography]
passlib
pytest
requests